Add GPU flag to render script
Currently, the only way to switch between CPU and GPU rendering is to
manually change the configuration in planets.blend file. Let's pass a
flag to the render script to do this programmatically from render
arguments.

For now, let's hardcode device type to METAL and GPU name to the test
GPU "AMD Radeon Pro 5500M".